+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Features
   + Java comments are generated as JavaDoc HTML and links to classes and members are supported now
   + Links to members in Swift comments are now supported
+  + Added support for structs with methods.
 
 ## 4.1.0
 - Release date: 2019-05-06